fix: import of moved types in vue
